refactor(cli): normalize for yarn/npm start
moved package manager helpers to a separate file, added normalization to post init setup
