+++ packages/amplify-cli/src/lib/init-steps/postInitSetup.js
@@ -1,6 +1,8 @@
 const { execSync } = require('child_process');
 const fs = require('fs');
 const path = require('path');
+const { getPackageManager } = require('../packageManagerHelpers')
+const { normalizePackageManagerForOS } = require('../packageManagerHelpers')
 
 async function run(context) {
   if (context.parameters.options.app) {
@@ -23,34 +25,17 @@ async function run(context) {
 }
 
 /**
- * Determine the package manager of the current project
- *
- * @return {string} 'yarn' if yarn.lock exists, 'npm' if package.json exists, undefined otherwise
-*/
-async function getPackageManager() {
-  const yarnLock = './yarn.lock';
-  const yarnLockDir = path.join(process.cwd(), yarnLock);
-  const packageJson = './package.json';
-  const packageJsonDir = path.join(process.cwd(), packageJson);
-  if (fs.existsSync(yarnLockDir)) {
-    return 'yarn';
-  } else if (fs.existsSync(packageJsonDir)) {
-    return 'npm';
-  }
-  return undefined;
-}
-
-/**
  * Install package using the correct package manager if package handling file exists
  *
  * @param packageManager either npm or yarn
  */
 async function runPackage() {
   const packageManager = getPackageManager();
+  const normalizedPackageManager = normalizePackageManagerForOS(packageManager);
   if (packageManager === 'yarn') {
-    await execSync('yarn start', { stdio: 'inherit' });
+    await execSync(`${normalizedPackageManager} start`, { stdio: 'inherit' });
   } else if (packageManager === 'npm') {
-    await execSync('npm start', { stdio: 'inherit' });
+    await execSync(`${normalizedPackageManager} start`, { stdio: 'inherit' });
   }
 }
